fix: ACPI0016 should use segment as UID
Currently the SSDT generation code is simply assigning incrementing
values to the UID field. This is a problem because the CEDT needs
to have the same UID values for the host bridges as the SSDT ACPI0016
has. To fix this, we change the PCD to use the segment number
as the UID. This should be knowable and the same for both
generators; segment numbers are known to be unique in our system.

Signed-off-by: Jeshua Smith <jeshuas@nvidia.com>
Reviewed-by: Ashish Singhal <ashishsingha@nvidia.com>
Reviewed-by: Vidya Sagar <vidyas@nvidia.com>
Tested-by: Jake Garver <jake@nvidia.com>
Reviewed-by: Jake Garver <jake@nvidia.com>